+++ CHANGELOG.md
@@ -8,7 +8,7 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
-## [3.1.3] - 2025-07-11
+## [3.1.3] - 2025-07-13
 
 ### Development Breaking Changes
 
@@ -27,6 +27,12 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [package]: Updated dependencies.
 - [vendorId]: Added Shortcut Labs Flic (0x1488).
+- [server]: Refactored serverNode event handlers and types.
+- [matterbridge]: Removed duplicated properties.
+
+### Fixed
+
+- [shutdown]: Fixed error messages from frontend when Matterbridge is shutting down.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">